Support stochastic (chance) constraints
We now assume constraints to be mappings X -> [0, 1], with the special case
that deterministic constraints always evaluate to 0 or 1. For VBMC, we discretize
the constraint probability by thresholding on 0.5, and for BOLFI, we sample Bernoulli rvs
with the corresponding constraint probabilities to support stochastic outcomes.

+++ harness/constraints/box.py
@@ -30,7 +30,7 @@ class BoxConstraint:
         self.b = jnp.array(b)
 
     def __call__(self, theta):
-        return jnp.all((self.a <= theta) & (theta <= self.b), axis=-1)
+        return jnp.float_(jnp.all((self.a <= theta) & (theta <= self.b), axis=-1))
 
     def __str__(self):
         formatted_constraints = []
